+++ presets/structured-cloning.js
@@ -15,6 +15,7 @@ import arraybuffer from '../types/arraybuffer.js';
 import typedArrays from '../types/typed-arrays.js';
 import dataview from '../types/dataview.js';
 import domexception from '../types/domexception.js';
+import domrect from '../types/domrect.js';
 
 import imagedata from '../types/imagedata.js';
 import imagebitmap from '../types/imagebitmap.js'; // Async return
@@ -64,6 +65,8 @@ const expObj = [
     /* c8 ignore next */
     typeof BigInt !== 'undefined' ? [bigint, bigintObject] : [],
     /* c8 ignore next */
-    typeof DOMException !== 'undefined' ? domexception : []
+    typeof DOMException !== 'undefined' ? domexception : [],
+    /* c8 ignore next */
+    typeof DOMRect !== 'undefined' ? domrect : []
 );
 export default expObj;