+++ test/test.js
@@ -1,6 +1,6 @@
 /* globals InternalError */
 /* globals document, ImageData, createImageBitmap, Blob, FileReader, File,
-    crypto, XMLHttpRequest, xmlHttpRequestOverrideMimeType */
+    crypto, DOMRect, XMLHttpRequest, xmlHttpRequestOverrideMimeType */
 /* eslint-disable no-restricted-syntax -- instanceof is
     convenient for checking here */
 /* eslint-disable new-cap -- For clarity */
@@ -49,6 +49,7 @@ const {
     undef, primitiveObjects, nan, infinity,
     negativeInfinity, date, error,
     regexp, map, set, arraybuffer, domexception,
+    domrect,
     dataview, imagedata, imagebitmap,
     blob, file, filelist, nonbuiltinIgnore,
     userObject, cloneable, resurrectable,
@@ -231,6 +232,27 @@ function Undefined (preset) {
         });
     });
 }
+
+/**
+ * @param {TypesonPreset} [preset]
+ * @returns {void}
+ */
+function DomRect (preset) {
+    describe('DOMRect', function () {
+        it('should return a DOMRect', function () {
+            const typeson = new Typeson().register(preset || [domrect]);
+            const domRect = new DOMRect(1, 2, 3, 4);
+            const tson = typeson.stringify(domRect, null, 2);
+            const back = typeson.parse(/** @type {string} */ (tson));
+            expect(back).to.be.an.instanceOf(DOMRect);
+            expect(back.x).to.equal(1);
+            expect(back.y).to.equal(2);
+            expect(back.width).to.equal(3);
+            expect(back.height).to.equal(4);
+        });
+    });
+}
+DomRect();
 
 /**
  * @param {TypesonPreset} [preset]
@@ -1638,6 +1660,7 @@ describe('Presets', () => {
         NonindexKeys(structuredCloningThrowing);
         CryptoKey(structuredCloningThrowing);
         DomException(structuredCloningThrowing);
+        DomRect(structuredCloningThrowing);
         it('should work with Structured cloning with throwing', () => {
             const typeson = new Typeson().register([structuredCloningThrowing]);
             expect(() => {